requirements.txt
apache-airflow==2.7.0
pymongo==4.6.0
psycopg2-binary==2.9.9
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
